% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_morans_i.R
\name{format.p}
\alias{format.p}
\title{Helper function for formatting p-values.}
\usage{
\method{format}{p}(p)
}
\arguments{
\item{p}{Numeric. A p-value.}
}
\value{
Character string. Formatted p-value.
}
\description{
Helper function for formatting p-values.
}
\keyword{internal}
