% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_data.R
\name{geocode_contamination_data}
\alias{geocode_contamination_data}
\title{Geocode Contamination Sample Addresses
This function takes a data frame with an address column and performs
geocoding using the 'osm' method from \code{tidygeocoder}. It then filters out
rows where geocoding failed (missing latitude/longitude) and converts
the resulting data into an \code{sf} (simple features) object for spatial analysis.}
\usage{
geocode_contamination_data(
  data,
  address_col,
  lat_col_name = "latitude",
  lon_col_name = "longitude",
  crs_value = 4269,
  verbose = FALSE
)
}
\arguments{
\item{data}{A data frame containing the addresses to be geocoded.
This should typically be the output from \code{clean_contamination_data()},
which already has a united address column.}

\item{address_col}{A character string specifying the name of the column
in \code{data} that contains the full addresses to be geocoded.}

\item{lat_col_name}{A character string specifying the desired name for the
latitude column in the geocoded output (default: "latitude").}

\item{lon_col_name}{A character string specifying the desired name for the
longitude column in the geocoded output (default: "longitude").}

\item{crs_value}{An integer or character string specifying the Coordinate
Reference System (CRS) for the output \code{sf} object (default: 4269, WGS84).}

\item{verbose}{Logical. If TRUE, \code{tidygeocoder} will print messages
during the geocoding process (default: FALSE).}
}
\value{
An \code{sf} object (simple features data frame) with geocoded coordinates
and filtered out rows where geocoding was unsuccessful. The geometry column
will be points representing the geocoded locations.
}
\description{
Geocode Contamination Sample Addresses
This function takes a data frame with an address column and performs
geocoding using the 'osm' method from \code{tidygeocoder}. It then filters out
rows where geocoding failed (missing latitude/longitude) and converts
the resulting data into an \code{sf} (simple features) object for spatial analysis.
}
